Apply theme on layout mounted.
