--- nuxt/entities/theme.ts
@@ -1,4 +1,0 @@
-export type Theme = 'light'|'dark'|'auto';
-
-export const THEME_PREFERENCE_STORAGE_KEY = 'theme:preference';
-export const THEME_APPLIED_STORAGE_KEY = 'theme:applied';
